chore(python): adopt ty type checking with a stub for the compiled core
Adds python/dataprof/_dataprof.pyi declaring the raw PyO3 extension symbols
so type checkers can resolve dataprof._dataprof (previously unresolvable).
The public __init__.pyi now re-exports these types instead of duplicating
them, and gains the missing ProfileReport.__init__.

- Wire ty (Astral, pre-1.0) into CI as an advisory, non-blocking step; add
  it to the dev dependency group and [tool.ty] config.
- Fix real findings surfaced en route: nullable config params, on_progress
  callback type, DataQualityMetrics Optional narrowing in tests, and the
  collections.abc import (UP035 under py310).
- ruff: enable isort combine-as-imports for tidy re-export blocks.

ty check, ruff, and pytest (139) all green.
